syscall: add open, write and flush syscall. write, read and lsee take s64 instead of s32

+++ include/levi/fs/file.h
@@ -12,13 +12,13 @@ s32 open(const char *pathname, u32 flags, proc_t *process);
 
 STATUS close(s32 fd, proc_t *prcess);
 
-s32 write(s32 fd, void *buffer, u32 size, proc_t *process);
+s64 write(s32 fd, void *buffer, u64 size, proc_t *process);
 
-s32 read(s32 fd, void *buffer, u32 size, proc_t *process);
+s64 read(s32 fd, void *buffer, u64 size, proc_t *process);
 
 s32 flush(s32 fd, proc_t *process);
 
-s32 lseek(s32 fd, u64 offset, u32 whence, proc_t *process);
+s64 lseek(s32 fd, u64 offset, u32 whence, proc_t *process);
 
 s32 dup2(s32 oldfd, s32 newfd, proc_t *process);
 
@@ -26,13 +26,13 @@ s32 kopen(const char *pathname, u32 flags);
 
 STATUS kclose(s32 fd);
 
-s32 kwrite(u32 fd, void *buffer, u32 size);
+s64 kwrite(u32 fd, void *buffer, u64 size);
 
-s32 kread(u32 fd, void *buffer, u32 size);
+s64 kread(u32 fd, void *buffer, u64 size);
 
 s32 kflush(u32 fd);
 
-s32 klseek(s32 fd, u64 offset, u32 whence);
+s64 klseek(s32 fd, u64 offset, u32 whence);
 
 s32 kdup2(s32 oldfd, s32 newfd);
