+++ kernel/drivers/framebuffer/framebuffer.c
@@ -5,10 +5,10 @@
 
 static file_t *__open(const char *name, u32 flags);
 static void __destroy_file(file_t *file);
-static s32 __write(file_t *file, u8 *buffer, u32 size);
-static s32 __read(file_t *file, u8 *buffer, u32 size);
+static s64 __write(file_t *file, u8 *buffer, u64 size);
+static s64 __read(file_t *file, u8 *buffer, u64 size);
 static s32 __flush(file_t *file);
-static s32 __lseek(file_t *file, u64 offset, u32 whence);
+static s64 __lseek(file_t *file, u64 offset, u32 whence);
 
 static struct vfs_operation framebuffer_operation = { .mkdir = NULL,
                                                       .rmdir = NULL,
@@ -124,7 +124,7 @@ static void __destroy_file(file_t *file)
     kfree(file);
 }
 
-static s32 __write(file_t *file, u8 *buffer, u32 size)
+static s64 __write(file_t *file, u8 *buffer, u64 size)
 {
     u32 i = 0;
 
@@ -138,7 +138,7 @@ static s32 __write(file_t *file, u8 *buffer, u32 size)
     return i;
 }
 
-static s32 __read(file_t *file, u8 *buffer, u32 size)
+static s64 __read(file_t *file, u8 *buffer, u64 size)
 {
     u32 i = 0;
 
@@ -162,7 +162,7 @@ static s32 __flush(file_t *file)
     return 0;
 }
 
-static s32 __lseek(file_t *file, u64 offset, u32 whence)
+static s64 __lseek(file_t *file, u64 offset, u32 whence)
 {
     switch (whence)
     {
